fix(header): isola dalle tecnologie assistive tutta la pagina dietro al menu mobile
Con il menu mobile aperto il plugin marcava come `inert` solo l'elemento `main`,
e solo quando la navbar era fuori da esso. Tutto il resto della pagina restava
raggiungibile dal cursore virtuale dei lettori di schermo: le fasce dello Slim
Header e dell'Header Centrale, il pulsante hamburger, header e footer del sito.
Il focus trap da solo non impedisce quel tipo di navigazione.

Il pannello ora usa la stessa isolazione già adottata dalle modali: risalendo la
catena degli antenati vengono marcati `inert` tutti i fratelli incontrati. Il
backdrop è escluso, perché `inert` blocca anche i puntatori e il click sul
backdrop deve continuare a chiudere il menu. L'attributo viene rimosso prima di
`hidden.bs.navbarcollapsible`, dato che il focus torna sul pulsante hamburger,
che è uno degli elementi resi inert.

La logica di `modal.js` è stata estratta in `util/inert.js` senza cambiarne il
comportamento: il conteggio dei riferimenti deve essere condiviso tra i
componenti, altrimenti due registri indipendenti rimuoverebbero `inert` da
elementi che l'altro sta ancora usando (es. una modale aperta dal menu mobile).

Refs #1753
